+++ mitra_config/src/federation.rs
@@ -5,7 +5,6 @@ const fn default_fetcher_timeout() -> u64 { 300 }
 const fn default_deliverer_timeout() -> u64 { 30 }
 const fn default_deliverer_log_response_length() -> usize { 75 }
 const fn default_fep_e232_enabled() -> bool { false }
-const fn default_fep_8b32_eddsa_enabled() -> bool { true }
 const fn default_announce_like_enabled() -> bool { true }
 
 #[derive(Clone, Deserialize)]
@@ -26,8 +25,6 @@ pub struct FederationConfig {
 
     #[serde(default = "default_fep_e232_enabled")]
     pub fep_e232_enabled: bool,
-    #[serde(default = "default_fep_8b32_eddsa_enabled")]
-    pub fep_8b32_eddsa_enabled: bool,
     #[serde(default = "default_announce_like_enabled")]
     pub announce_like_enabled: bool,
 }
@@ -43,7 +40,6 @@ impl Default for FederationConfig {
             onion_proxy_url: None,
             i2p_proxy_url: None,
             fep_e232_enabled: default_fep_e232_enabled(),
-            fep_8b32_eddsa_enabled: default_fep_8b32_eddsa_enabled(),
             announce_like_enabled: default_announce_like_enabled(),
         }
     }