+++ CHANGELOG.md
@@ -29,6 +29,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Deprecate OpenSSL support.
 
+### Removed
+
+- Removed `federation.fep_8b32_eddsa_enabled` configuration parameter.
+
 ## [2.21.0] - 2024-06-01
 
 ### Added